+++ gulp/tasks/cleanup.js
@@ -3,7 +3,7 @@ import logger from "gulplog";
 
 import { delConfig } from "../config";
 
-export const cleanup = done =>
+export const cleanup = (done) =>
   del(delConfig).then(() => {
     logger.info("Folders `build` and `tmp` were successfully deleted");
     done();