+++ gulp/tasks/development.js
@@ -4,4 +4,4 @@ import { server } from "./server";
 import { build } from "./build";
 import { watch } from "./watch";
 
-export const dev = gulp.series(build, gulp.parallel(server, watch));
+export const development = gulp.series(build, gulp.parallel(server, watch));