+++ gulp/tasks/serviceWorker.js
@@ -1,10 +1,10 @@
 import swPrecache from "sw-precache";
 import logger from "gulplog";
 
-import { isDevelopment } from "../util/env";
+import { isDevelopment } from "../util/environment";
 import swConfig from "../../sw-precache-config";
 
-export const serviceWorker = done => {
+export const serviceWorker = (done) => {
   swConfig.logger = logger.info;
   swConfig.handleFetch = !isDevelopment;