+++ gulpfile.babel.js
@@ -1,14 +1,14 @@
 import gulp from "gulp";
 import { build } from "./gulp/tasks/build";
 import { zip } from "./gulp/tasks/zip";
-import { dev } from "./gulp/tasks/dev";
+import { development } from "./gulp/tasks/development";
 import { deploy } from "./gulp/tasks/deploy";
 import { major, minor, patch, reset } from "./gulp/tasks/semver";
 
 // Main tasks
 gulp.task("build", build);
 gulp.task("zip", zip);
-gulp.task("dev", dev);
+gulp.task("development", development);
 gulp.task("deploy", deploy);
 
 // Semver
@@ -18,4 +18,4 @@ gulp.task("semver:patch", patch);
 gulp.task("semver:reset", reset);
 
 // Default task
-gulp.task("default", gulp.series("dev"));
+gulp.task("default", gulp.series("development"));