+++ source/scripts/helpers/dom/getSiblings.js
@@ -4,19 +4,17 @@
  * @param  {Node}  elem The element
  * @return {Array}      The siblings
  */
-let getSiblings = function(elem) {
-  let siblings = [];
-  let sibling = elem.parentNode.firstChild;
+const getSiblings = function (element) {
+  const siblings = [];
+  let sibling = element.parentNode.firstChild;
+
   for (; sibling; sibling = sibling.nextSibling) {
-    if (sibling.nodeType === 1 && sibling !== elem) {
+    if (sibling.nodeType === 1 && sibling !== element) {
       siblings.push(sibling);
     }
   }
+
   return siblings;
 };
 
 export default getSiblings;
-
-// Usage example:
-// let elem = document.querySelector("#some-element");
-// let siblings = getSiblings(elem);