+++ source/scripts/helpers/dom/toArray.js
@@ -1,3 +1,3 @@
-const nodeListToArray = nodeList => [].slice.call(nodeList);
+const nodeListToArray = (nodeList) => [].slice.call(nodeList);
 
 export default nodeListToArray;