+++ sw-precache-config.js
@@ -7,16 +7,18 @@ const patterns = [
   `${dist}/**/*.html`,
   `${dist}/**/*.js`,
   `${dist}/**/*.css`,
-  `${dist}/assets/**`
+  `${dist}/assets/**`,
 ];
 
 const paths = [];
-patterns.forEach(pattern =>
+patterns.forEach((pattern) =>
   paths.push(...glob.sync(pattern.replace(path.sep, "/")))
 );
 
 module.exports = {
   navigateFallback: "/index.html",
   stripPrefix: `${dist}/`,
-  staticFileGlobs: paths.filter(path => !/assets\/images\/content/.test(path))
+  staticFileGlobs: paths.filter(
+    (path) => !/assets\/images\/content/.test(path)
+  ),
 };