+++ CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to
 [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [9.6.0] - 2021-01-23
+
+### Fixed
+
+- Updated dependencies and fix vulnerabilities
+- Terser config in webpack config
+- Fithub actions workflow config
+- Eslint config
+
+### Changed
+
+- Renamed npm command `npm run amo` to `npm run new`
+
 ## [9.5.1] - 2019-04-08
 
 Use `npm` in travis config
@@ -137,7 +150,7 @@ Use `npm` in travis config
 ### Changed
 
 - All javascript plugins rewrited using Plugin class
-- Updated javascript template for `amo` command
+- Updated javascript template for `npm run new` command
 - Updated to new html-like api checkbox, radio, fieldset, form-item, select, textarea, avatar, dialog, modal, spinner, bage, dropdown, accordion components
 - Comonent bread-crumbs has new api and new design
 - New button style `text`